+++ jawsome-cli/Changelog.md
@@ -1,3 +1,45 @@
+# 1.2.0
+
+jawsome-cli is now extensible with arbitrary custom functions!
+
+### Passing custom values in
+
+Pass through custom arguments at the command line using `-X key:val`.
+
+e.g. If I have a custom value I want to pass in to my denorm phase, I
+would use `java -jar path-to-my-uberjar.jar denorm -X
+custom-key:custom-val`. Note that `custom-key` will be parsed as
+"everything before the first `:`" so `custom-key` may not contain a `:`.
+
+### Using the custom values on the other side
+
+`def-cli-pipeline` is now variadic -- it takes an l2-form AND any
+number of functions to invoke at runtime. These "cli-time-thunks"
+should generally look like this:
+
+    (defn set-custom-val! [xs]
+      (let [cli-custom-val (get xs "custom-val")]
+        (swap! custom-val (constantly cli-custom-val))))
+
+That is, they take a map of all the xs that were passed in through
+`-X`, get some value in the map of xs, and execute some code with
+side-effects.
+
+You can then dethunk the thunk in a pipeline like so:
+
+    (def custom-val (atom nil))
+    (defvar 'yield-custom-val-map
+      (fn []
+        {"custom-val" @custom-val}))
+
+    (def-cli-pipeline
+      '(pipeline
+         (denorm-phase
+           (xform-phase (custom :static-values (dethunk (ref yield-custom-val-map))))))
+      set-custom-val!)
+
+Verbose but it gets the job done.
+
 # 1.1.0
 
 Added `--field-order` option for schema-phase, `--record-terminator`